fix: preserve canvas for LED matrix
All[^1] of the hardware bugs I thought I had... were just software
bugs.

The offscreen canvas is _created_ at each invocation to
`offscreen_canvas`; while it's owned by the matrix, it isn't singular
(apparently). We have to swap in the canvas that we wrote to, not just
"an" offscreen canvas.

(Does this mean that the previous version was leaking canvases all over
the place? Or just destroying the one in place? I don't know -- I
haven't checked the C++ code.)

This restores the swapin/swapout behavior from the *working* demo.

[^1]: Not _all_ -- this wouldn't cover the overlaying of red on half of the
display. Still don't know what that was about