fix: for i2c access, set up i2c before matrix library
I wasn't getting a reliable startup either (a) running as root (no
access to i2c device), or (b) running as non-root (no access to /dev/mem
for fast GPIO). Why?

I ran the daemon under strace to see.
When running as root, something was calling `setresuid`
and `setresgid` to switch from `root` (UID 0) to `daemon` (UID 1),
which didn't have access to /dev/i2c-1.

The fix seems to be opening the i2c device before starting the matrix
library. That way we have an I2c file descriptor open.

Ultimately, I'll need to work out the /dev/mem permissions to run this
in userspace; but, whatever, that's a later problem.